requirements.txt
autopep8
click
colorama
Flask==2.1.3
Flask-Login
Flask-MySQLdb
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
mysqlclient
pycodestyle
toml
Werkzeug==2.0.3
WTForms
